correction for 1-print_binary

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -8,7 +8,7 @@
 
 void print_binary(unsigned long int n)
 {
-	unsigned int mask = 1;
+	unsigned long int mask = 1;
 
 	while ((mask << 1) <= n)
 	{
